commit ; travaul sur ke style

+++ TFE_GestionDeStockEtVenteEnLigne/Models/Metier/Produit.cs
@@ -19,8 +19,8 @@ namespace TFE_GestionDeStockEtVenteEnLigne.Models
 
         public String Denomination { get; set; }
 
-        [Range(0, float.MaxValue)]
-        public float Prix { get; set; }
+        [Range(0, double.MaxValue)]
+        public double Prix { get; set; }
 
         [Range(0, int.MaxValue)]
         public int QuantiteEmballage { get; set; }
